Adding SourceInfo interface (#39)
* Add SourceInfo interface to force implementing validate() in all SourceInfo classes
* Rename SourceInfo abstract class and interface names
